Restructure database controller files
